+++ src/pyedb/dotnet/edb_core/cell/primitive/__init__.py
@@ -1,0 +1,3 @@
+from pathlib import Path
+
+workdir = Path(__file__).parent
