+++ src/pyedb/dotnet/edb_core/edb_data/padstacks_data.py
@@ -26,7 +26,7 @@ import re
 import warnings
 
 from pyedb.dotnet.clr_module import String
-from pyedb.dotnet.edb_core.cell.primitive import Primitive
+from pyedb.dotnet.edb_core.cell.primitive.primitive import Primitive
 from pyedb.dotnet.edb_core.dotnet.database import PolygonDataDotNet
 from pyedb.dotnet.edb_core.edb_data.edbvalue import EdbValue
 from pyedb.dotnet.edb_core.general import PadGeometryTpe, convert_py_list_to_net_list
