+++ src/pyedb/dotnet/edb_core/nets.py
@@ -206,7 +206,7 @@ class EdbNets(object):
         for net in self._layout.nets[:]:
             total_plane_area = 0.0
             total_trace_area = 0.0
-            for primitive in net.Primitives:
+            for primitive in net.primitives:
                 if primitive.GetPrimitiveType() == self._edb.cell.primitive.PrimitiveType.Bondwire:
                     continue
                 if primitive.GetPrimitiveType() != self._edb.cell.primitive.PrimitiveType.Path:
@@ -1088,7 +1088,7 @@ class EdbNets(object):
         else:
             if not start_with and not contain and not end_with:
                 net = self._edb.cell.net.find_by_name(self._active_layout, net_name)
-                if net.IsNull():
+                if net.is_null:
                     net = self._edb.cell.net.create(self._active_layout, net_name)
                 return net
             elif start_with:
